Record relation accesses only for reference tables
When we implemented foreign keys to reference tables, we first implemented
a relation access tracking infrastructure (#2220). That was generic enough
to follow all the relation accesses.

However, for the purposes of foreign keys to reference tables, we only need
to track relation accesses to reference tables.

This change especially important because the relation access tracking is
very inefficient for partitioned tables with many partitions. So, we decided
to only record relevant tables, which are reference tables only.

(Note that we're already tracking the Parallel relation accesses via
RecordParallelRelationAccessForTaskList() once per distributed query.)