use transactions to create dependencies on nodes
The commands that are sent to the worker to create dependencies are prepended with a command to disable ddl propagation. This fails when not done in a transaction for installations that have a pg_bouncer setup between the nodes. By sending all the commands in a transaction they will reuse the same pg_bouncer connection and therefore are ran with ddl propagation	disabled.