Allow join order planner to not output a plan
In some cases it is better to not output a plan than to fall back to a cartesian product. Once the cartesion product is used in a plan the error message becomes non-understandable and the cartesian product is not implemented anyway.

Instead we only output the cartesian product if there are no applicable join clauses between tables, eg. all tuples on the left need to be multiplies by the tuples on the right without any filter. This will still give the cartesian product error for queries with tables that have no join clauses, but fails to omit a plan with the cartesian product if there is a non-supported join clause on a table. This allows us to tell the user complex queries are only supported in certain situations.