Skip parse/deparse of local fast-path queries
If we're dealing with local fast-path queries, we can skip the expensive
operations of parse/deparse of the queries.

Given that the planner cannot know in advance whether the query is going
to be executed locally, it tries it's best, skips expensive operations.

However, the executor could end-up with doing the expensive operations
if needed.