Lazily generate deparsed queries
Deparsing and parsing a query can be heavy on CPU. In some cases we
don't need to do this in theory, such as when locally executing the
query.

This commit is the first step in allowing to skip deparsing and parsing
the query in these cases, by lazily creating the query string and
storing the query in the task. Future commits will make use of this and
not deparse and parse the query anymore, but use the one from the task
directly.