Register placement accesses in local execution
After a local execution, we tried to continue locally executing queries
to make sure all previous changes are visible. This commit relaxes this
constraint by checking if any placements are accessed during local execution.

If no placements are accessed in a locally executed subplan, the successive
subplans are not forced to run locally.